Calendar day select functionality bug fixed
